Grade This Commit
Please grade this commit. 
This commit will not compile on my machine because I could not get the threading libraries to work in Eclipse IDE. I probably spent more time trying to get the libraries to work than I actually did on this phase of the project. (I would continually get "thread is not declared in this scope" despite having the proper libraries and compile settings.) 
In theory I believe this commit should compile and work correctly. However, if it doesn't please grade the commit directly AFTER this one. That commit will not have multithreading implemented, but has all other features of Phase II.
If it is necessary that the project must compile, please grade the